# internal/service/resources/api/openapi.yaml
...
openapi: "3.0.0"
info:
  title: O2IMS Infrastructure Inventory API
  version: 1.0.0
  description: |
    API for O2IMS Infrastructure Inventory
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
...
servers:
- url: http://localhost:8000
...
paths:
  /o2ims-infrastructureInventory/api_versions:
    get:
      operationId: getAllVersions
      summary: Get API versions
      description: |
        Returns the complete list of API versions implemented by the service.
      tags:
      - metadata
      responses:
        '200':
          description: |
            Successfully obtained the complete list of versions.
          content:
            application/json:
              schema:
                $ref: "../../common/api/openapi.yaml#/components/schemas/APIVersions"
        '400':
          description: Bad request
          content:
            application/problem+json:
              schema:
                $ref: '../../common/api/openapi.yaml#/components/schemas/ProblemDetails'
        '500':
          description: Internal server error.
          content:
            application/problem+json:
              schema:
                $ref: '../../common/api/openapi.yaml#/components/schemas/ProblemDetails'
...
  /o2ims-infrastructureInventory/v1/api_versions:
    get:
      operationId: getMinorVersions
      summary: Get minor API versions
      description: |
        Returns the list of minor API versions implemented for this major version of the API.
      tags:
      - metadata
      responses:
        '200':
          description: |
            Success
          content:
            application/json:
              schema:
                $ref: "../../common/api/openapi.yaml#/components/schemas/APIVersions"
        '400':
          description: Bad request
          content:
            application/problem+json:
              schema:
                $ref: '../../common/api/openapi.yaml#/components/schemas/ProblemDetails'
        '500':
          description: Internal server error.
          content:
            application/problem+json:
              schema:
                $ref: '../../common/api/openapi.yaml#/components/schemas/ProblemDetails'
...